# Reduce heap allocation, remove unnecessary mut, derive Clone