This improves usability and marginally improves performance:

Before:
```
test sha256_10 ... bench:       4,176 ns/iter (+/- 88) = 2 MB/s
test sha256_1k ... bench:      68,458 ns/iter (+/- 2,005) = 14 MB/s
test sha256_8k ... bench:     520,298 ns/iter (+/- 14,572) = 15 MB/s
```

After:
```
test sha256_10 ... bench:       4,113 ns/iter (+/- 535) = 2 MB/s
test sha256_1k ... bench:      66,892 ns/iter (+/- 8,179) = 15 MB/s
test sha256_8k ... bench:     506,935 ns/iter (+/- 57,681) = 15 MB/s
```