@Ralith 2018-03-14
Added a separate commit that changes the API to an [RFC5869](https://tools.ietf.org/html/rfc5869)-inspired one, which I find more intuitive when implementing IETF standards, especially because the argument order differs. If you'd prefer to pass on that, I'll back it out.