@warner 2018-03-14
First commit seems totally fine to me.

Matching the RFC function names makes sense, but if we're thinking about moving this under the RustCrypto umbrella (#6) then we might want to line up with their KDF API.

@newpavlov , any thoughts on this? Looks like we should circle back to https://github.com/RustCrypto/traits/issues/5 and come to a conclusion.